things

+++ sample-bot.py
@@ -155,12 +155,12 @@ def buyNormalStocks(exchange):
         bids = data['buy']
 
         for price, size in bids:
-            if price > fvList[sym][0]:
+            if price > (int)fvList[sym][0]:
                 trades.append(('SELL', sym, price, size))
 
         asks = data['sell']
         for price, size in asks:
-            if price < fvList[sym][0]:
+            if price < (int)fvList[sym][0]:
                 trades.append(('BUY', sym, price, size))
     return trades
